+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.6.5 – unreleased
+### Added
+- Add helper view for keyboard shortcuts
+  [#91](https://github.com/nextcloud/mail/pull/91) @Gomez 
+
+### Changed
+- TODO
+
+### Fixed
 - TODO
 
 ## 0.6.4 – 2017-05-02